Clean up GraphQL schema code
