+++ graphqlapi/src/main/java/graphql/kickstart/servlet/SchemaGraphQLServlet.java
@@ -19,9 +19,9 @@ import graphql.kickstart.execution.GraphQLObjectMapper;
 import graphql.schema.GraphQLSchema;
 import io.stargate.auth.AuthenticationService;
 import io.stargate.db.Persistence;
-import io.stargate.graphql.core.KeyspaceManagementSchema;
 import io.stargate.graphql.graphqlservlet.CassandraUnboxingGraphqlErrorHandler;
 import io.stargate.graphql.graphqlservlet.GraphqlCustomContextBuilder;
+import io.stargate.graphql.schema.SchemaFactory;
 
 public class SchemaGraphQLServlet extends SimpleGraphQLHttpServlet {
   private final Persistence persistence;
@@ -45,6 +45,6 @@ public class SchemaGraphQLServlet extends SimpleGraphQLHttpServlet {
   }
 
   private GraphQLSchema createSchema() {
-    return new KeyspaceManagementSchema(persistence, authenticationService).build().build();
+    return SchemaFactory.newDdlSchema(persistence, authenticationService);
   }
 }
