+++ graphqlapi/src/main/java/io/stargate/graphql/schema/fetchers/CassandraFetcher.java
@@ -1,0 +1,40 @@
+package io.stargate.graphql.schema.fetchers;
+
+import graphql.schema.DataFetcher;
+import graphql.schema.DataFetchingEnvironment;
+import io.stargate.auth.AuthenticationService;
+import io.stargate.auth.StoredCredentials;
+import io.stargate.db.ClientState;
+import io.stargate.db.Persistence;
+import io.stargate.db.QueryState;
+import io.stargate.db.datastore.DataStore;
+import io.stargate.graphql.graphqlservlet.HTTPAwareContextImpl;
+
+/** Base class for fetchers that access the Cassandra backend. It also handles authentication. */
+public abstract class CassandraFetcher<ResultT> implements DataFetcher<ResultT> {
+
+  protected final Persistence<?, ?, ?> persistence;
+  protected final AuthenticationService authenticationService;
+
+  public CassandraFetcher(
+      Persistence<?, ?, ?> persistence, AuthenticationService authenticationService) {
+    this.persistence = persistence;
+    this.authenticationService = authenticationService;
+  }
+
+  @Override
+  public final ResultT get(DataFetchingEnvironment environment) throws Exception {
+    HTTPAwareContextImpl httpAwareContext = environment.getContext();
+
+    String token = httpAwareContext.getAuthToken();
+    StoredCredentials storedCredentials = authenticationService.validateToken(token);
+    ClientState clientState = persistence.newClientState(storedCredentials.getRoleName());
+    QueryState queryState = persistence.newQueryState(clientState);
+    DataStore dataStore = persistence.newDataStore(queryState, null);
+
+    return get(environment, dataStore);
+  }
+
+  protected abstract ResultT get(DataFetchingEnvironment environment, DataStore dataStore)
+      throws Exception;
+}
