+++ graphqlapi/src/main/java/io/stargate/graphql/schema/fetchers/ddl/CreateTableFetcher.java
@@ -13,57 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.stargate.graphql.fetchers;
+package io.stargate.graphql.schema.fetchers.ddl;
 
 import com.datastax.oss.driver.api.core.metadata.schema.ClusteringOrder;
 import com.datastax.oss.driver.api.querybuilder.SchemaBuilder;
 import com.datastax.oss.driver.api.querybuilder.schema.CreateTable;
 import com.datastax.oss.driver.api.querybuilder.schema.CreateTableStart;
 import com.datastax.oss.driver.api.querybuilder.schema.CreateTableWithOptions;
-import graphql.schema.DataFetcher;
 import graphql.schema.DataFetchingEnvironment;
 import io.stargate.auth.AuthenticationService;
-import io.stargate.auth.StoredCredentials;
-import io.stargate.db.ClientState;
 import io.stargate.db.Persistence;
-import io.stargate.db.QueryState;
-import io.stargate.db.datastore.DataStore;
-import io.stargate.db.datastore.ResultSet;
-import io.stargate.graphql.graphqlservlet.HTTPAwareContextImpl;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
 
-public class CreateTableDataFetcher
-    implements io.stargate.graphql.fetchers.SchemaFetcher, DataFetcher {
-  private final Persistence persistence;
-  private AuthenticationService authenticationService;
+public class CreateTableFetcher extends DdlQueryFetcher {
 
-  public CreateTableDataFetcher(
-      Persistence persistence, AuthenticationService authenticationService) {
-    this.persistence = persistence;
-    this.authenticationService = authenticationService;
-  }
-
-  @Override
-  public Object get(DataFetchingEnvironment environment) throws Exception {
-    HTTPAwareContextImpl httpAwareContext = environment.getContext();
-
-    String token = httpAwareContext.getAuthToken();
-    StoredCredentials storedCredentials = authenticationService.validateToken(token);
-    ClientState clientState = persistence.newClientState(storedCredentials.getRoleName());
-    QueryState queryState = persistence.newQueryState(clientState);
-    DataStore dataStore = persistence.newDataStore(queryState, null);
-
-    CompletableFuture<ResultSet> resultSetSingle = dataStore.query(getQuery(environment));
-    resultSetSingle.get();
-    return true;
+  public CreateTableFetcher(
+      Persistence<?, ?, ?> persistence, AuthenticationService authenticationService) {
+    super(persistence, authenticationService);
   }
 
   public String getQuery(DataFetchingEnvironment dataFetchingEnvironment) {
     CreateTableStart start =
         SchemaBuilder.createTable(
-            (String) dataFetchingEnvironment.getArgument("keyspaceName"),
+            dataFetchingEnvironment.getArgument("keyspaceName"),
             (String) dataFetchingEnvironment.getArgument("tableName"));
 
     Boolean ifNotExists = dataFetchingEnvironment.getArgument("ifNotExists");
