+++ graphqlapi/src/main/java/io/stargate/graphql/util/CaseUtil.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.stargate.graphql.core;
+package io.stargate.graphql.util;
 
 import com.google.common.collect.ImmutableMap;
 import java.util.Map;
