+++ graphqlapi/src/test/java/io/stargate/graphql/schema/DdlTestBase.java
@@ -1,0 +1,9 @@
+package io.stargate.graphql.schema;
+
+import graphql.schema.GraphQLSchema;
+
+public abstract class DdlTestBase extends GraphQlTestBase {
+  protected GraphQLSchema createGraphQlSchema() {
+    return SchemaFactory.newDdlSchema(persistence, authenticationService);
+  }
+}
